refactoring, moving icons to assets folder
