--- src/apps/pulse/components/Close.tsx
@@ -1,24 +1,0 @@
-export default function Close(props: { onClose: () => void }) {
-  const { onClose } = props;
-  return (
-    <button
-      style={{ height: 20, width: 20, marginRight: 10 }}
-      onClick={() => onClose()}
-      type="button"
-      aria-label="Close"
-    >
-      <svg
-        xmlns="http://www.w3.org/2000/svg"
-        viewBox="0 0 20 20"
-        fill="currentColor"
-        className="size-5"
-      >
-        <path
-          fillRule="evenodd"
-          d="M10 18a8 8 0 1 0 0-16 8 8 0 0 0 0 16ZM8.28 7.22a.75.75 0 0 0-1.06 1.06L8.94 10l-1.72 1.72a.75.75 0 1 0 1.06 1.06L10 11.06l1.72 1.72a.75.75 0 1 0 1.06-1.06L11.06 10l1.72-1.72a.75.75 0 0 0-1.06-1.06L10 8.94 8.28 7.22Z"
-          clipRule="evenodd"
-        />
-      </svg>
-    </button>
-  );
-}
