--- src/apps/pulse/components/Esc.tsx
@@ -1,37 +1,0 @@
-interface CloseProps {
-  closePreview: () => void;
-}
-
-export default function Esc(props: CloseProps) {
-  const { closePreview } = props;
-  return (
-    <button
-      style={{
-        backgroundColor: 'black',
-        borderRadius: 10,
-        width: 40,
-        height: 40,
-      }}
-      onClick={() => {
-        closePreview();
-      }}
-      type="button"
-      aria-label="Save"
-    >
-      <svg
-        width="36"
-        height="34"
-        viewBox="0 0 36 34"
-        fill="none"
-        xmlns="http://www.w3.org/2000/svg"
-      >
-        <rect width="36" height="34" rx="8" fill="#1E1D24" />
-        <path
-          opacity="0.5"
-          d="M8.35811 14.161V16.813H11.4781V18.022H8.35811V20.791H11.8681V22H6.87611V12.952H11.8681V14.161H8.35811ZM16.4105 22.091C15.8038 22.091 15.2578 21.987 14.7725 21.779C14.2872 21.5623 13.9058 21.259 13.6285 20.869C13.3512 20.479 13.2125 20.024 13.2125 19.504H14.7985C14.8332 19.894 14.9848 20.2147 15.2535 20.466C15.5308 20.7173 15.9165 20.843 16.4105 20.843C16.9218 20.843 17.3205 20.7217 17.6065 20.479C17.8925 20.2277 18.0355 19.907 18.0355 19.517C18.0355 19.2137 17.9445 18.9667 17.7625 18.776C17.5892 18.5853 17.3682 18.438 17.0995 18.334C16.8395 18.23 16.4755 18.1173 16.0075 17.996C15.4182 17.84 14.9372 17.684 14.5645 17.528C14.2005 17.3633 13.8885 17.112 13.6285 16.774C13.3685 16.436 13.2385 15.9853 13.2385 15.422C13.2385 14.902 13.3685 14.447 13.6285 14.057C13.8885 13.667 14.2525 13.368 14.7205 13.16C15.1885 12.952 15.7302 12.848 16.3455 12.848C17.2208 12.848 17.9358 13.069 18.4905 13.511C19.0538 13.9443 19.3658 14.5423 19.4265 15.305H17.7885C17.7625 14.9757 17.6065 14.694 17.3205 14.46C17.0345 14.226 16.6575 14.109 16.1895 14.109C15.7648 14.109 15.4182 14.2173 15.1495 14.434C14.8808 14.6507 14.7465 14.9627 14.7465 15.37C14.7465 15.6473 14.8288 15.877 14.9935 16.059C15.1668 16.2323 15.3835 16.371 15.6435 16.475C15.9035 16.579 16.2588 16.6917 16.7095 16.813C17.3075 16.9777 17.7928 17.1423 18.1655 17.307C18.5468 17.4717 18.8675 17.7273 19.1275 18.074C19.3962 18.412 19.5305 18.867 19.5305 19.439C19.5305 19.8983 19.4048 20.3317 19.1535 20.739C18.9108 21.1463 18.5512 21.4757 18.0745 21.727C17.6065 21.9697 17.0518 22.091 16.4105 22.091ZM20.5252 17.463C20.5252 16.579 20.7289 15.786 21.1362 15.084C21.5522 14.382 22.1112 13.836 22.8132 13.446C23.5239 13.0473 24.2995 12.848 25.1402 12.848C26.1022 12.848 26.9559 13.0863 27.7012 13.563C28.4552 14.031 29.0012 14.6983 29.3392 15.565H27.5582C27.3242 15.0883 26.9992 14.733 26.5832 14.499C26.1672 14.265 25.6862 14.148 25.1402 14.148C24.5422 14.148 24.0092 14.2823 23.5412 14.551C23.0732 14.8197 22.7049 15.2053 22.4362 15.708C22.1762 16.2107 22.0462 16.7957 22.0462 17.463C22.0462 18.1303 22.1762 18.7153 22.4362 19.218C22.7049 19.7207 23.0732 20.1107 23.5412 20.388C24.0092 20.6567 24.5422 20.791 25.1402 20.791C25.6862 20.791 26.1672 20.674 26.5832 20.44C26.9992 20.206 27.3242 19.8507 27.5582 19.374H29.3392C29.0012 20.2407 28.4552 20.908 27.7012 21.376C26.9559 21.844 26.1022 22.078 25.1402 22.078C24.2909 22.078 23.5152 21.883 22.8132 21.493C22.1112 21.0943 21.5522 20.544 21.1362 19.842C20.7289 19.14 20.5252 18.347 20.5252 17.463Z"
-          fill="white"
-        />
-      </svg>
-    </button>
-  );
-}
