+++ src/apps/pulse/components/Search/Search.tsx
@@ -9,21 +9,22 @@ import React, {
 import { TailSpin } from 'react-loader-spinner';
 import { useLocation, useNavigate } from 'react-router-dom';
 import { isAddress } from 'viem';
-import { useTokenSearch } from '../hooks/useTokenSearch';
-import Close from './Close';
+import { useTokenSearch } from '../../hooks/useTokenSearch';
+import Close from '../Misc/Close';
 import {
   Asset,
   parseFreshAndTrendingTokens,
   parseSearchData,
-} from '../utils/parseSearchData';
-import { chainNameToChainIdTokensData } from '../../../services/tokensData';
-import { SearchType, SelectedToken } from '../types/tokens';
-import Refresh from './Refresh';
+} from '../../utils/parseSearchData';
+import { chainNameToChainIdTokensData } from '../../../../services/tokensData';
+import { SearchType, SelectedToken } from '../../types/tokens';
+import Refresh from '../Misc/Refresh';
 import ChainSelectButton from './ChainSelect';
-import { getChainId, MobulaChainNames } from '../utils/constants';
+import { getChainId, MobulaChainNames } from '../../utils/constants';
 import ChainOverlay from './ChainOverlay';
 import TokenList from './TokenList';
-import { isTestnet } from '../../../utils/blockchain';
+import { isTestnet } from '../../../../utils/blockchain';
+import SearchIcon from '../../assets/seach-icon.svg';
 
 interface SearchProps {
   setSearching: Dispatch<SetStateAction<boolean>>;
@@ -180,19 +181,7 @@ export default function Search(props: SearchProps) {
             }}
           >
             <span style={{ marginLeft: 10 }}>
-              <svg
-                width="20"
-                height="20"
-                viewBox="0 0 24 24"
-                fill="none"
-                stroke="currentColor"
-                strokeWidth="2"
-                strokeLinecap="round"
-                strokeLinejoin="round"
-              >
-                <circle cx="11" cy="11" r="8" />
-                <path d="m21 21-4.3-4.3" />
-              </svg>
+              <img src={SearchIcon} alt="search-icon" />
             </span>
             <input
               ref={inputRef}
